+++ skoolkit/sna2skool.py
@@ -76,6 +76,8 @@ def main(args):
                        help="Specify a control file to use, or a directory from which to read control files. "
                             "PATH may be '-' for standard input, or '0' to use no control file. "
                             "This option may be used multiple times.")
+    group.add_argument('-C', '--comments', action='store_true',
+                       help="Generate instruction comments.")
     group.add_argument('-d', '--defb', dest='defb', metavar='SIZE', type=int,
                        help='Disassemble as DEFB statements of this size.')
     group.add_argument('-e', '--end', dest='end', metavar='ADDR', type=integer, default=65536,