+++ sphinx/source/changelog.rst
@@ -5,8 +5,8 @@ Changelog
 -----
 * Dropped support for Python 3.8
 * Added the :ref:`comment generator <commentGenerator>` component
-* Added the ``--comments`` option to :ref:`sna2ctl.py` (for generating
-  instruction comments)
+* Added the ``--comments`` option to :ref:`sna2ctl.py` and :ref:`sna2skool.py`
+  (for generating instruction comments)
 * Added the ``Comments`` configuration parameter for
   :ref:`sna2ctl.py <sna2ctl-conf>` (to enable instruction comment generation)
 * :ref:`trace.py` now responds to keypresses while running with screen contents